Casting long int to "MONTH".
Sonarqube:
http://nemo.sonarqube.org/issues/search#issues=9be82142-884b-43c1-b9b0-3b848dbee6df

+++ examples/src/main/java/org/apache/mahout/classifier/NewsgroupHelper.java
@@ -56,7 +56,7 @@ public final class NewsgroupHelper {
   public static final int FEATURES = 10000;
   // 1997-01-15 00:01:00 GMT
   private static final long DATE_REFERENCE = 853286460;
-  private static final long MONTH = 30 * 24 * 3600;
+  private static final long MONTH = (long)30 * 24 * 3600;
   private static final long WEEK = 7 * 24 * 3600;
   
   private final Random rand = RandomUtils.getRandom();
